Fixed GCC warnings: -Wconversion -Wshadow -Wunused-parameter. JIRA:CXX-8058
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@72358 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/util/test/test_uttp.cpp
@@ -275,7 +275,7 @@ bool CUTTPWriterTest::Test(size_t buffer_size)
     m_Writer.Reset(buffer, buffer_size);
     m_WriterTestStream = s_TestStream;
 
-    int counter = sizeof(s_TestChunkSequence) / sizeof(*s_TestChunkSequence);
+    size_t counter = sizeof(s_TestChunkSequence) / sizeof(*s_TestChunkSequence);
 
     const SChunkData* test_seq = s_TestChunkSequence;
